requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-oauth-toolkit>=2.3.0
django-guardian>=2.4.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0 
django-cors-headers>=4.7.0